splits styles back to be in their own directories
