+++ app/javascript/mastodon/features/account_timeline/v2/styles.module.scss
@@ -1,0 +1,36 @@
+.filtersWrapper {
+  padding: 16px 24px 8px;
+}
+
+.filterSelectButton {
+  appearance: none;
+  border: none;
+  background: none;
+  padding: 8px 0;
+  font-weight: 500;
+  display: flex;
+  align-items: center;
+}
+
+.filterSelectIcon {
+  width: 16px;
+  height: 16px;
+}
+
+.filterOverlay {
+  background: var(--color-bg-primary);
+  border-radius: 12px;
+  box-shadow: var(--dropdown-shadow);
+  min-width: 230px;
+  display: grid;
+  grid-template-columns: 1fr auto;
+  align-items: stretch;
+  row-gap: 16px;
+  padding: 8px 12px;
+
+  > label {
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+  }
+}
